# docs/contracts/openapi.yaml
...
openapi: 3.0.0
paths:
...
  /files/sign:
    post:
      operationId: FilesController_signUpload
      summary: Issue signed upload params (admin only) — cloudinary | local
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUploadDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUploadParamsDto'
      tags:
        - files
  /files/upload:
    post:
      operationId: FilesController_upload
      summary: Upload file vào local volume (admin only, STORAGE_DRIVER=local)
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFileDto'
      responses:
        '200':
          description: ''
      tags:
        - files
  /files/{id}:
    delete:
      operationId: FilesController_remove
      summary: Delete file (admin only) — DB record + revoke storage asset
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - files
...
info:
  title: MyBlog API
  description: REST + WebSocket contract — auto-generated from NestJS Swagger
  version: 0.2.0-alpha
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local dev
  - url: https://myblog-api.fly.dev
    description: Production
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    SignUploadDto:
      type: object
      properties:
        folder:
          type: string
          example: myblog/posts
          description: 'Cloudinary folder (default: myblog)'
        publicId:
          type: string
          example: post-abc-image-1
          description: Optional public_id override
        resourceType:
          type: string
          enum:
            - image
            - raw
          example: image
          description: image cho PNG/JPG/WebP, raw cho PDF/DOC/...
      required:
        - resourceType
    SignedUploadParamsDto:
      type: object
      properties:
        provider:
          type: string
          enum:
            - cloudinary
            - local
          example: cloudinary
        signature:
          type: string
          example: abc123signaturehash
        timestamp:
          type: number
          example: 1715952000
        apiKey:
          type: string
          example: '123456789012345'
        cloudName:
          type: string
          example: my-cloud
        folder:
          type: string
          example: myblog/posts
        resourceType:
          type: string
          enum:
            - image
            - raw
          example: image
        publicId:
          type: object
          example: post-abc-image-1
          nullable: true
        uploadUrl:
          type: string
          example: /files/upload
          description: 'Local driver: endpoint upload multipart'
      required:
        - provider
        - signature
        - timestamp
        - apiKey
        - cloudName
        - folder
        - resourceType
        - publicId
    UploadFileDto:
      type: object
      properties:
        folder:
          type: string
          example: myblog/posts
        resourceType:
          type: string
          enum:
            - image
            - raw
          example: image
        publicId:
          type: string
          example: u-abc-1717000000
      required:
        - folder
        - resourceType
...